Fix updating Team name failing
This commits fixes a bug where the Team name was
not being updated successfully. The root cause
was that the Team class was trying to create a
new team object instead of updating the existing one.

This commit modifies the Team class to check for
an existing team first with the passed in id
and only create a new one if it does not exist.